// src/model/incidentCreateBody.ts
/**
 * API BHB Customer Backend
 * API BHB Customer Backend
 *
 * OpenAPI spec version: 1.0.6
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Incident information
 */
export interface IncidentCreateBody { 
    /**
     * Incident comments
     */
    comments?: string;
    /**
     * Incident fault
     */
    incidentType?: IncidentCreateBody.IncidentTypeEnum;
}
export namespace IncidentCreateBody {
    export type IncidentTypeEnum = 'Kitchen' | 'Delivery' | 'Client';
    export const IncidentTypeEnum = {
        Kitchen: 'Kitchen' as IncidentTypeEnum,
        Delivery: 'Delivery' as IncidentTypeEnum,
        Client: 'Client' as IncidentTypeEnum
    };
}